% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission.R
\name{comission_calc}
\alias{comission_calc}
\title{Calculates product and sales comission.}
\usage{
comission_calc(perf_path, targets_path)
}
\arguments{
\item{perf_path}{The path to staff performance report}

\item{targets_path}{The path to targets files}
}
\value{
Dataframe with comission calculated.
}
\description{
Calculates product and sales comission.
}
\examples{
comission_calc('perf.csv','targets.csv')
}
