% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission.R
\name{product_calc}
\alias{product_calc}
\title{Calculates product comission based on sales.
0 - 300 %5
301 to 750 10%
751 and aobove 15%}
\usage{
product_calc(product_sales)
}
\arguments{
\item{product_sales}{A number. Product sales}
}
\value{
The comission on the product sales.
}
\description{
Calculates product comission based on sales.
0 - 300 %5
301 to 750 10%
751 and aobove 15%
}
\examples{
product_calc(100.00)
}
